Ignore uninitialized warning in cryptonote_core ctor
`m_blockchain_storage` is indeed uninitialized when used here during
construction, but we don't actually touch it until after it has been
initialized.

This points at real initialization problems in the code base, but not
something we're going to fix right now (and not a bug in the current
implementation).