Move/replace type_to_hex and hex_to_type
These were very similar to view_guts, aside from hex conversion; this
unifies the interfaces into guts.h which is now responsible for both
type-to-hex and hex-to-type conversions alongside type-to-byte and
byte-to-hex conversions.

The new methods are also more robust, and can view (or accept) `unsigned
char` or `std::byte` strings as well.

- make_from_hex_guts<T>(hex) -- constructs a T from its hex content
- hex_guts(x) -- converts an `x` to hex; replaces `type_to_hex`
- load_from_hex_guts(hex, x) -- loads `hex` into `x`; throws on error
- try_load_from_hex_guts(hex, x) -- loads `hex` into `x`; returns true
  on success, false on error.