Fix get_transaction_pool endpoint
The missing declaration in the header meant that the overload for
GET_TRANSACTION_POOL was never called (but instead just the base class
GET_TRANSACTIONS was invoked), and so `request.memory_pool = true;`
didn't happen, and thus the endpoint would always return an empty list.

This fixes it, and also removes some commented out code and an unused
argument.