Make periodic_task exception safe with error logging
If an exception is thrown the task does not get counted as completed and
will be tried again the next time it is invoked.

This also simplifies the random delay to be a single value because
having two values for the delay pointlessly adds a degree of freedom:
the delay distribution `a + U[b, c]` is equivalent to `(a+b) + U[0,
c-b]`, and so you can just give the base (a+b) and max random offset
(c-b) to achieve the exact same distribution in two parameters instead
of three.