Update ethyl for async; cpr via ethyl; updates for newer cpr
This also updates ethyl to the latest working PR with async support for
upcoming async commits.

It also gets cpr via ethyl instead of direct submodule because our quite
old cpr submodule commit was breaking ethyl (because it deferred to
ours).  Getting it via cpr gets rid of a submodule and isn't something
that we change very often in core, so ethyl's version will do fine.

This commit includes some small fixes for cpr usage in core code that
was broken by the newer cpr version ethyl brings in.