Clean up blockchain init and raw pointer usage
Blockchain::init was overcomplicated, with lots of code for utilities
and tests having to pass a bunch of nullptr arguments.  This cleans it
up with two separate init functions that greatly reduces the nullptrs
that have to be passed through.

The documentation for `init` also got enhanced here to be clear about
what the ownership implications of the various pointer arguments are.

It also cleans up the BlockchainDB ownership to contain it in a
unique_ptr instead of a raw pointer to make the ownership model clear.