Update BLS signing to match contract specific signing scheme
We switched to a scheme that hashes the data with Keccak256 instead of
SHA2, but additionally instead of expanding the input message to 256
bytes in Solidity we expand the message to 96 bytes before trying to map
the bytes onto BN256.

In order to recreate the same construction on both C++ and Solidity, C++
has to conform to the Solidity implementation. This commit adds those
necessary changes.