Fix BLS sig verification by using our new method from eth-sn-contracts
I've updated the way we generate and aggregate signatures to match our
smart contracts in eth-sn-contracts. The contracts use a new method of
`H(m||i) => (x, b)` which was raised by the Zellic audit such that it matches
the original BLS paper construction.

This required changes to herumi/bls as the implementation there
deviated. In doing that change we no longer use the sign/verify
functions provided by BLS but instead replace those calls with our own
ones that use the lower-level BLS operations of herumi/bls to achieve
the original construction.

In Jason's branch he went ahead and hashed the message contents
together with keccak. This is _currently_ not being done in the smart contracts.
This difference was causing signature verification to fail. For now
since the smart contracts are frozen I have reverted this codebase to
construct with the raw components in the binary buffer for `m` in `H(m||i)`
and sign that payload instead so that the smart contracts are able to verify
the signatures the C++ Session Nodes produce.