Optimization + avoid deprecation warnings
Static local variables are relatively expensive (because the compiler
must put a mutex or other thread-safe locking construct around them to
ensure that they are only assigned once); this moves the value outside
of the function instead (and also marks it constexpr).

It also de-compounds some assignments to avoid C++20 volatile
deprecation errors.  (Essentially: C++20 deprecates `x += value` for a
volatile `x` because it has often been misued as an atomic operation,
when in fact it is not).