Flag duplicate-non-standard tx failures
On devnet I see a lot of error logs such as:

    Transaction verification failed: 9d3ac3038d7e64bf02680835a0d55ad832699430c42905c277b7d88584a6343b
    Transaction verification failed: b840ffbf1daeb097e0a002b7b7ef8100cd48650cc008dca1366326b3ff68af66

that occur when a quorum state change tx arrives that duplicates one we
already have (but without being a perfect duplicate: typically because
it has an equivalent but different set of signatures).

This isn't an error, though, and we shouldn't be error logging when it
happens.  This adds a `m_duplicate_nonstandard` flag to the transaction
verification context so that we can detect such a case and just produce
a debug, instead of error, log message.

I couldn't find this in mainnet logs (looking at a dozen nodes), but
that might just be that the quorum vs network size is much smaller on
mainnet, and so this is rare -- I can't see any reason it *couldn't*
also produce such a spurious error log on mainnet.