Disable (by default) building rarely used utilities
The blockchain utilities binaries, in particular, are rarely needed, as
is the trusted multisig wallet generator; disabling them by default
saves a bit of compilation (and especially linking) time.

This also adds a BUILD_EVERYTHING option so that (for devs) a
`-DBUILD_EVERYTHING=ON` will turn all the optional bits on to make sure
they are still building.