Fix syncing with pruned reward rate data
Syncing a node was failing because it wasn't handling the "don't have
reward data" case from the L2 tracker (because the L2 tracker only keeps
the reward rate for recent heights).

This changes it to return an optional (instead of 0) for such a case,
and handles it properly.

It also adds a block production failure if trying to construct a block
without the reward available from the L2Tracker.