Fixes for unexpected data in print_pool
The print_pool command, when the pool contains a registration, errors
with something like:

    9b6af07917a22080eee1ff3ed9b834b2fceb736519ccbc1b64f5573b5f7f156d:
        size/weight: 227/227
        fee: 0.000000000 (0.000000000/byte)
        received: 1720581795 (11.9 hours ago)
        status: not relayed, do not relay
    Command errored: print_pool, [json.exception.type_error.302] type must be number, but is number

when it is trying to print the transaction; the error is because the
next line here is the "max height dependence" line, which was throwing
because max_used_height isn't set in the RPC response when 0, but the
command executor was assuming it was always set.  This fixes it, along
with a closely related weirdness in the blockchain code:

`max_used_block_id` was being set to the id of block 0 (i.e. the root
block) when max_used_block_height=0 (which it will be for non-spending
transactions such as registrations), but that is supposed to mean "there
is no height dependence", not that it depends specifically on block 0;
this updates it to set max_used_block_id to null hash rather than the
first blockchain block in such a case.

(The rest of the changes here are clang-formatting)