Improve x25519 lookup at eth HF
Previously we couldn't have the x25519 map in the state_t because the
X25519 pubkey is derived from the service node's Ed25519, which we
didn't learn about until we got an uptime proof with it.

In HF21 we now require unified pubkeys, and so we can compute it
instantly without needing to wait for a proof.

This implements it by keeping the x25519 pubkey map in the state_t and
updating it on initialization, and when service nodes are added/removed
so that we always know the full set of x25519 pubkeys without having to
wait for proofs.

Doing this resolves some quorumnet communication failures (such as in
pulse participation) that can happen when service nodes try working with
other service nodes but the proof of the connecting SN hasn't propagated
yet (or was missed).

This also changes the OMQ SN list code to include all SNs instead of
only active SNs with proofs: decommed SNs can still sometimes try to
communicate over oxenmq with other nodes and there's no reason they need
to be excluded from such comms.  (Pre-HF21 we still can only include
nodes with proofs, of course, because without a proof we don't know the
X25519 pubkey that will be used).