Don't print stake data by default; update URLs
Currently the prepare_eth_registration has a lot of output (from the
various keys and signatures), which isn't needed in the common case of
going via the staking website.

This changes the `print_only` option to just `print`: if given then we
just print it, if not given then we just submit it (but don't print it).

Also updates the URLs from the placeholders to the live staking website
URLs.