Bump version & fix timestamp in snapshot debs
Snapshot debs were using a ~devYYYMMDDHHMMSS in the version, but the MM
was using minutes instead of the month.  This fixes it (and also changes
to use UTC), but also bumps the main oxen version to 11.0.1 so that we
don't end up with an "older" package from the corrected date value.