Remove non-operator contributor minimum staking requirements for ETH regs
Enforcement of minimum stakes now happens in the multi-contributor
contract.  There are, however, cases in which the old strictly ordered
Oxen rules might not be satisfied (e.g. if a large contributor withdraws
from an in-progress open stake), but a full set of contributions can
*always* be rearranged to satisfy the rule.  But not enforcing it
Oxen-side we can save a little trouble of needing the multi contributor
contract to worry about sorting the contributor list before submitting
it to the SN contract.