Add map serialization support
Maps serialization works with the generic container + pair serialization
code with some small tweaks.

The main tweak is that we need to be able to specify the (otherwise
deduced) Container::value_type because the one for pair has a read-only
key, but we need the mutable key alternative for insertion.

Also includes a fix for the `.insert` code path not actually working.