rpc: only put block_header or block_headers in output if requested
If you only use it in single hash or multi-hash mode then you should
only see the single header or array of headers output.  Currently in
single hash mode it was always including an empty block_headers array
which was confusing.

Also shortens the `get_block_header_by_hash` name to `gbh` because the
previous name was too verbose (and cumbersome to type) for a variable
with such a short scope.