Add block sync speed to Synced messages
Before:

    Synced 413816/584100  (70%, 170284 left)

After:

    Synced 413816/584100  (70%, 170284 left)  (616 blocks/sec)

The speed is quite useful to see to get an idea of how long syncing will
take, but is only shown if you put oxend into info+ verbosity, along
with a bunch of other less useful info.  This adds the speed to the
message always (but keeps the other stuff at info+ or debug+ verbosity).