Allow shutting down during subsystem rescan
Currently oxend's subsystem rescan is not abortable, and so if you try
killing or restarting oxend, it simply won't acknoledge the signal.
On service nodes, this often results in systemd timing out the stop and
restoring to a SIGKILL, which is not a nice way to go (and can cause
lmdb corruption or bloating if done at the wrong time).

This commit weaves an atomic<bool> through the potentially slow parts of
core and blockchain initialization so that when we *do* want to stop
(from Ctrl-C, or a regular TERM signal) we gracefully fail and exit
rather than needing a `kill -9` to actually stop.