Optimization: skip verification during SN state rescan
When we rescan the service node state from the blockchain (which is
sometimes necessary after a crash, or after certain pop_block patterns)
we do an unnecessary full pulse signature verification of the block, but
this is unnecessary on this path: the fact that we added the block to
the LMDB at some point means it already passed the required checks.

This roughly doubles the SN state rescan speed (because of no longer
needing to extract and verify signatures on each pulse block).

This also lowers the number of blocks we process at once to 100 instead
of 500 as we sometimes have reports that 1) it can respond (to a
shutdown) faster, and 2) on overloaded servers processing 500 blocks can
take so long that it triggers the systemd watchdog (because our ping of
systemd is itself in this loop).  The reduction to 100 didn't appear to
affect performance at all on my (unloaded) system so seems reasonable to
do.