Fix mempool selection/drop penalty zone under SENT
It was trying to apply Oxen reward values for SENT blocks during mempool
calculation (in order to compute when adding a tx starts reducing the
total reward), but that won't work in SENT because there *are* no Oxen
reward values under SENT to reduce.

Instead this updates the mempool selection algorithm to simply make the
300kB previously "soft" block limit a hard limit starting at HF21,
dropping support for the (rarely used) dynamic mempool size and penalty
zone.