p2p: replace stupid "parabola \/" selection with truncated exponential
The peer selection random value here was braindead: it would return 0
45% of the time, would return 1-5, 7, 9, 11, 13, 16, 19 5% of the time,
and other indices never.

This was dumb.  This commit replaces it with a truncated exponential,
tuned to return values 0-9 75% of the time (see added comment for other
distribution examples), and removes the arbitrary peer selection limit
of 20 (which was actually 19, because of the terrible implementation
described above).