Make liquidation RPC take a ed25519 key for consistency
In most of our public facing interfaces the identity of a node is
represented by their Ed25519 key following the BLS hardfork. This
updates the RPC call for aggregation to accept an Ed25519, do a lookup
for the BLS key and then aggregate a signature on it.

This had a knock-on effect and exposed a bunch of bugs in the
implementation which were somewhat addressed in this patch.

- With exits implemented, funds are now returned to the staker. However
we don't have migration code yet so the zero-eth-address was assigned to
the contributors. This led the SQL DB to insert a cryptonote address as
the destination address to return the stake to. This eventually stalled
the network as it could not parse the address out of the DB.

There needs to be more safe-guards for this and insert into the DB
should fail if this is to happen instead of silently succeeding. I've
added a check in user code to block the insertion of the zero address
but this probably needs to be more robust by having the check at the DB layer.

- It's now possible for a SN to be present in the smart contract but not
the SNL yet because of vote confirmations. These pending registrations
are now accounted for when calculating if a node is eligible to be
removed. Before that, these nodes could be prematurely ejected from the
smart contract, their registrations could then be confirmed and desync
the SNL and smart contract.

Since this was even remotely possible at one point, it's likely that the
SNL and smart contract can desync in the future. Further mitigations are
needed here to force them back to sync if this is to happen.

- Started setting up code paths for allowing BLS aggregation from
decommissioned nodes and nodes that have recently been removed (either
the user requested a voluntary leave and the unlock height was met, or,
the node was deregistered). We want as many nodes to participate in BLS
aggregation because, whilst their node is removed from the SNL they are
_not_ removed from the smart contract yet.

Since they are not removed from the smart contract yet, their BLS keys
are still contributing to the master BLS key on the smart contract. If
they're still contactable then their participation in the signing of the
BLS signature to flush the list of their presence increases the
stability of the network.

However this doesn't work yet as it seems like the unauthenticated
connections to these nodes never get a response and time-out.