Default initialised JSON archiver doesn't serialize
The stack_ tracking the depth in which you are in the JSON object that
you are serializing was always empty when default initialised. In the
set function it checks if the stack is empty and if it's empty it
overwrites the entire payload with the value.

This means when serialising an entire struct it constantly overwrites
the entire JSON object with the field leaving you with an incorrect
serialisation of the object.

This patch fixes that by removing that check. Since there's always
atleast one object globally at the root in the JSON that is returned when we
call `curr()`, we can use that object to fill out the object and hence
serialise the entire struct correctly.