Use top-block when initially evaluating Pulse round
We were previously using multiple individuals calls to retrieve
information about the top block. Though unlikely, it's possible for the
top block to change inbetween the individual invocation, for example an
alt-block could come in and change the operating assumptions Pulse
should be making.

The first step here is to use info from the snapped copy of the top
block and extrapolate information from there. This is atomic and so we
use consistent data to initialise the Pulse rounde.