In Pulse use top_hash to determine if the chain has changed
Previously we used the chain-height. This has issues when we reorg to
the same height. The height remains the same however the top block has
changed. When the top block changes, all our operating assumptions about
how Pulse is to be conducted changes because the entropy is derived from
the top block.

This meant that when you re-orged, the node would still attempt to use
old block information that is no longer available and cause Pulse to
fail for that node prematurely. Switching to the top hash means this is
detected immediately and it can reconfigure accordingly.