Fix SQL reset_database, fix delayed payments conflict
You can have multiple delayed payments (e.g. queued stakes) to be
returned back to the user after they exit their node if the operator has
exited multiple of their nodes or nodes they contribute to on the same
block. SQL was erroring out with a conflict on the primary key of the
table (address, payout_height) which can clearly conflict.

We sum up the amounts to be returned if there are more than 1 stake
being returned for the payout height. Not to mention the payout height,
although unlikely could conflict with a voluntary exit that coincides
with a deregistration by chance.

Additionally, reset_database was not dropping the delayed_payments
table, nor, setting the height back to 0.