Replace herumi with AntelopeIO/bn256
Herumi code is an extremely efficient implementation, but it poses
numerous problems for distribution:

- it doesn't build on arm with gcc
- it doesn't build with older gcc (e.g. the one in Ubuntu 20.04) on x86-64
- building it with clang on x86-64 results in a binary that aborts with
  an illegal instruction during static initialization when not used on
  an avx512 CPU.
- CPU instruction detection fails in some environments and causes an
  illegal instruction even when built with gcc.
- The package doesn't build properly, and upstream frequently accepts
  patches without understanding them that makes the build *worse* (such
  as a recent one to link everything to both static *and* dynamic builds
  of the library at the same time).

The illegal instructions are caused by its (invalid) build forcing
`-mavx512` into the compilation, which it needs to be able to get at
some avx512 primitives.  This breaks, however, because this also enables
problematic avx instruction use in the code *outside* the cpu detection
guarded areas.

On top of those, the code itself is a mess that has pursued speed above
maintainability for years, and it shows.

I've given up trying to get it to work and instead went a C++ port of
geth's go implementation.

It is slower than Herumi, primarily centered around the hot path
functions in src/bitint_arithmetic.h.  I also investigated alternative
bigint implementations here:
    - gmp was slower
    - https://github.com/piggypiggy/fp256 was the same speed
    - a hand rolled 256-bit mul implementation was the same speed

It isn't unusable slow, however; on my system (Ryzen 5950X):
    - signing takes ~0.4ms
    - signature verification takes ~2.6ms
    - secretkey -> pubkey takes ~0.1ms
    - pubkey aggregation is very fast (1000 pks aggregate in ~0.8ms)
    - signature aggregation is slow (1000 sigs aggregate in 390ms)

Those numbers are a bit less than half the speed of the same operations
with Herumi (relative speeds between the operations are pretty much the
same).

I've extensively tested this with many thousands of randomly generated
keys and cross-validation of this code versus a herumi build to ensure
that we get exactly the same pubkeys, signatures, and verification
values across both versions.  (The pk/signature values being unchanged
in the test suite from the ones we get with Herumi, for instance, are an
example of this).

Note: this switches to our fork of the upstream project to incorporate
some small warning fixes and support for 32-bit targets.