Drop boost::uuids::uuid because boost 1.85+ broke it
`boost::uuids::uuid` used to be a plain-old-data type; boost 1.85 broke
it so that it is neither plain old data, nor constexpr, and that breaks
all sorts of things internally where we expect those things, such as its
use in network_config, and the ability to directly serialize it on the
wire as POD data.

This rips it out and replaces it with a simple `epee::connection_id_t`
(which is basically just a type-safe `std::array<unsigned char, 16>`)
which works exactly the same but without the upstream breakage.