Don't require rebuilding everything from turning unit tests on/off
The -DUNIT_TEST added here means we have to rebuild the entire code base
if we turn unit tests on (or off) from an existing build.  This removes
it by allowing the handful of affected functions to be non-static
always, but only exposes the definitions in the one test code .cpp that
wants it.