Add separate beneficiary to ethereum payouts
This change affects consensus by adding a new beneficiary field to the
contributors of an Ethereum contributor in the
`eth::event::NewServiceNode` which is a `tx_extra` field which gets
hashed and hence must be agreed upon by all nodes on the network to be
added to the blockchain.

All existing nodes will have their beneficiary defaulted to the wallet
that staked the contribution hence it is backwards compatible for other
parts of the code to use the new beneficiary field without a hardfork
such as the SNL and SQL DB.

The SNL will recalculate the metadata to ensure that all old nodes will
have their beneficiary set to the wallet's address.

When we intend to hardfork, we will pause the smart contracts, upgrade
them and keep the contracts paused until the network has transitioned
over. This ensures all nodes have upgraded and can parse the new
contract payloads.