Resolve HF consensus on New SN problem
Instead of trying to retrofit the NewServiceNode emitted log to store
both a V1 and V2 payload, simply introduce a new V2 log that is handled in
a separate step from the V1 payload.

All nodes will observe and agree that a log is V2 when the contract is
upgraded. The contract is _not_ upgraded until all nodes on the network
have hardforked. Thus nodes will always be in agreement of the canonical
representation of the log.