Add support for serialization of std::optionals
For json serialization these become either the value, or the json `null`
literal.

For binary serialization these become a serialized bool (0 or 1)
indicating whether the value is set and then, if (and only if) 1, the
serialized value.  Deserializing loads the bool and then either clears
the optional (if 0) or emplaces and loads it (if 1).