Remove unneeded lock
This `lock` parameter has long confused me.  On looking into it, the
removed `lock` bool is used in only one place, but it doesn't appear to
do anything even there: m_db->get_height() is already thread-safe, and
so any locking mechanics that would be needed here have to happen
*outside* the `get_current_blockchain_height()` method as needed, not
inside it.

Also remove the big scary WARNING in the function body (and others)
because it's really not warning about *anything important*: yes, without
a lock, the return value of `get_current_blockchain_height()` (or
whatever) might be stale when you return it.  But that's basically true
of *anything* multithreaded so the warning seems pointless and probably
worse counterproductive because it makes it seem like there's something
special here in this regard when there isn't.