thread-safety for recently-removed node access
The `recently_removed_nodes()` method was returning a span into the
internal nodes, but that's not thread safe and could break if a new
block arrived while iterating through it and updated the sn state.

This changes the interface to a callback-taking
`for_each_recently_removed_node` that iterates through the nodes while
holding the needed SN list lock.