Check event vote status in mempool selection
While testing the recent commits, I added a temporary hack that
deliberately voted against all ServiceNodeExits as I wanted to force a
oxen/contract mismatch.  This ended up stalling the chain for a couple
hours because all service nodes had the event, and block proposers would
include it, even though they would vote against it.

This isn't correct behaviour: the block producer, by proposing a block,
is implicitly saying that they are in favour of all events within that
block.  While we shouldn't add a state change to the mempool in the
first place that we aren't in favour of, there are potential ways that
such a transaction could still get there (such as maliciously relayed
events, or reorgs) and so we should always be checking that we actually
would vote in favour of any event that we include.

This seems a subtle edge case, and it isn't entirely clear to me that
such a case is triggerable without such a hack, but the extra check is
simple and can't hurt anything.

This commit extends the purge-specific vote check code (which explicitly
needs such a check) to be a generic vote check of any event prior to
inclusion.