Don't send proofs if L2 provider non-responsive (HF20+)
If we haven't heard from the L2 provider since more than an hour ago (or
whatever the regular uptime proof interval is) then don't send proofs so
that the problem gets noticed.

This also adds a temporary --l2-skip-proof-check command/config option
that can be used during HF20 to bypass this check during the transition
period, in case operators run into L2 configuration problems that take a
while to solve.