DRY/fix height setting actions
Height can get set on two different paths: either when fetching the
height from all nodes for a are-we-in-sync check, or otherwise just from
the primary as a regular update.  The first path was missing a call to
prune_old_states and so if you had the sync check set low enough to run
on every update the old states would never get pruned.

This DRYs the code to do the height appropriately in a single method and
fix the discrepency.