Fix delayed_payments table on pop_blocks / detach
The delayed_payments table in the SQLite db was not getting correctly
updated when blocks were removed (either by popping or
blockchain_detached).  This adds a trigger on height change similar to
the batched payments archive which simply removes entries added
after/above the new height.

Also simplifies a query to use MAX.