Fix sqlite db height decreased triggers
In the sqlite db, archive height X is for block height X - 1, so when we
revert/decrease to height X we want to keep the archive for that height,
as it is still valid (and likely what we just restored from).
Previously it was removing this archive and subsequent detach/pop calls
would have to look further back in the archive for no reason.