Use new stagenet seed node address
Attempting to reuse the previous seed node address across the stagenet
reboot causes problems: the (new) seed node will ban unupgraded nodes,
and then when they upgrade, they can't connect to the seed node and thus
can't find any peers.

This switches the seed node to use a port that was never a peer on the
old network and so should be "clean" for the rebooted stagenet.