Avoid L2 provider requests for SN exit signing
This replaces the handling for determining if a node is removable to
avoid needing to go to the L2 tracker, and thus eliminate a potential L2
provider credit exhaustion attack where a malicious entity could request
signatures over and over to make a service node exhaust its available L2
provider request credits.  This instead now uses a combination of the
recently_removed_nodes list and the infrequently-updated purge list
(that was added for removing not-in-contract nodes) to handle the
removability determination without needing any L2 requests other than
the ones already being done to maintain chain state.

This change slightly alters the definition of what we consider
"removable":

- recently_removed_nodes are now instantly considered removable.

- nodes in the contract and not in oxend are considered removable, but
  with a lag because we now use the cached service node list collected
  for node purging (which is updated only once an hour or so).

This avoids any need for an L2 request when signing; "normal" signatures
will be for the first case, and won't incur any delay.  The second case
is mainly of a fallback for abnormal handling (for instance: a
registration accepted by the smart contract that oxend rejected for some
reason), where an hour or more of delay doesn't meaningfully matter.

Other related changes made here:

- shorten the mainnet purge list interval to 1h instead of 2h as it now
  makes sense to be moderately more up-to-date now that it's also used
  for missing nodes in both directions instead of just contract-missing
      nodes.

- for_each_pending_l2_state now allows a bool lambda to return true to
  short-circuit the iteration (similar to how some other for_each_...
  work).

- Remove unused `get_all_bls_public_keys` method; it seems everything
  that might have used it is using `get_all_service_node_ids` instead.

- Added a more efficient, singleton "is removable" method that doesn't
  need to produce to full removable list for the common case of a check
  for a single removable node (as is common for signing requests to
  SNs).

- Make get_removable_nodes return an unordered_map<pubkey, bool> instead
  of a vector as it simplifies both implementation and common uses (the
  key presence indicates removability, the bool indicates liquidatable).

- Consolidate is-removable/liquidatable functions into blockchain
  (removing from cryptonote_core).

- Re-enable is_removable/is_liquidatable fields in get_service_nodes rpc
  response.