Store {block,tx,contrib} index in delayed_payments
We do a dupe insertion check against the {block,tx,contrib} index which
is guaranteed to be unique. Note we do _not_ check/store block hash
because our threat model here is that the block hash is valid by the
time this part of the code is met (e.g. it passed all of blockchain and
SNL validation essentially).

To do this we have to pipe through some new information from the SNL
when it calls into SQL DB to insert a delayed payment (i.e. return of
stake). This makes it so that essentially there is 1 row for each
returned stake as they are uniquely identified by the {block,tx,contrib}
index.