Use HISTORY_KEEP_RECENT_WINDOW in SNL&SQL to unify stored backups
Incremental steps to getting the the SQL and SNL to handle
blockchain_detached in the same manner. They both need to detach to
exactly the same height so they must also store backups to the same
degree on both sides so that one subsystem doesn't desync on detach.

If they desync, we risk the 2 systems producing inconsistent rewards
which breaks BLS aggregation.