Use special local-devnet deploy script to avoid exit-too-early by 2hrs problem
The devnet version of the script deploys a contract that inherits the
mainnet ServiceNodeRewards contract but instead overrides the minimum
exit time. The contract currently rejects exits that occur within the
first 2hrs of the node registering.

We don't have an easy way to get the nodes to fast-fwd their clocks by
2hrs to sign a signature a time in the future.