Fix bool return in add_sn_rewards being unused in db_sqlite
This function always returns true meaning downstream checks on this call
always succeed. This function can fail at the SQL insertion step.
Predominantly, the pattern in the SQL DB is to use exceptions as the
error mechanism and the one function that does call `reward_handler` is
wrapped in a try-catch block.

I've opted to remove the boolean return and instead annotate that the
function throws. This simplifies the calling code in that it no longer
has to do a success check (which was redundant since true was always
returned) and instead optimistically executes and defers to the
try-catch to handle errors.