Fix broken 'make_archive'
Make archive was calculating the wrong height to prune a DB by when
a block is processed and it's a height eligible for archiving rows. It
calculates

```
height < (NEW.HEIGHT - HISTORY_ARCHIVE_KEEP_WINDOW) % HISTORY_ARCHIVE_INTERVAL
```

The RHS calculation is always going to produce a height in the range
[0, HISTORY_ARCHIVE_INTERVAL) which is not going to give us a rolling window of
the past HISTORY_ARCHIVE_KEEP_WINDOW archives. I've updated this to

```
height < (NEW.HEIGHT - HISTORY_ARCHIVE_KEEP_WINDOW)
```

There might be some archives that slip through if the KEEP_WINDOW and
ARCHIVE_INTERVAL are not divisible by each other but this should not be
a big deal.