Fix pulse_sorter.last_height_validating_in_quorum off-by-1
The height that is meant to be applied to the sort keys for selecting
the Pulse quorum is meant to be the _height_ of the block that is being
processed. When some of the code was shuffled around to implement the TX
confirmation approach the incrementing of the SNL's height was done after
instead of before the quorum was generated.

This the nodes to be included in a quorum to be subtly off-by-1 once the
sorting goes out of sync.